GH-683: interim commit
