+++ node/src/accountant/db_access_objects/failed_payable_dao.rs
@@ -4,7 +4,7 @@ use crate::accountant::db_access_objects::utils::{
 };
 use crate::accountant::db_big_integer::big_int_divider::BigIntDivider;
 use crate::accountant::{checked_conversion, comma_joined_stringifiable};
-use crate::blockchain::errors::{AppRpcError, BlockchainDbError, PreviousAttempts};
+use crate::blockchain::errors::validation_status::PreviousAttempts;
 use crate::database::rusqlite_wrappers::ConnectionWrapper;
 use itertools::Itertools;
 use masq_lib::utils::ExpectValue;
@@ -381,9 +381,9 @@ mod tests {
         make_read_only_db_connection, FailedTxBuilder,
     };
     use crate::accountant::db_access_objects::utils::current_unix_timestamp;
-    use crate::blockchain::errors::{
-        AppRpcError, AppRpcErrorKind, LocalError, PreviousAttempts, RemoteError,
-        ValidationFailureClockReal,
+    use crate::blockchain::errors::blockchain_db_error::app_rpc_web3_error_kind::AppRpcWeb3ErrorKind;
+    use crate::blockchain::errors::validation_status::{
+        PreviousAttempts, ValidationFailureClockReal,
     };
     use crate::blockchain::test_utils::{make_tx_hash, ValidationFailureClockMock};
     use crate::database::db_initializer::{
@@ -592,7 +592,7 @@ mod tests {
         assert_eq!(
             FailureReason::from_str(r#"{"Submission":{"Local":{"Decoder"}}}"#).unwrap(),
             FailureReason::Submission(PreviousAttempts::new(
-                Box::new(AppRpcErrorKind::Decoder),
+                Box::new(AppRpcWeb3ErrorKind::Decoder),
                 &validation_failure_clock
             ))
         );
@@ -643,7 +643,7 @@ mod tests {
 
         assert_eq!(
             FailureStatus::from_str(r#"{"RecheckRequired":{"Reattempting":{"ServerUnreachable":{"firstSeen":{"secs_since_epoch":1755080031,"nanos_since_epoch":612180914},"attempts":1}}}}"#).unwrap(),
-            FailureStatus::RecheckRequired(ValidationStatus::Reattempting( PreviousAttempts::new(Box::new(AppRpcErrorKind::ServerUnreachable), &validation_failure_clock)))
+            FailureStatus::RecheckRequired(ValidationStatus::Reattempting( PreviousAttempts::new(Box::new(AppRpcWeb3ErrorKind::ServerUnreachable), &validation_failure_clock)))
         );
 
         assert_eq!(
@@ -726,7 +726,7 @@ mod tests {
             .reason(PendingTooLong)
             .status(RecheckRequired(ValidationStatus::Reattempting(
                 PreviousAttempts::new(
-                    Box::new(AppRpcErrorKind::ServerUnreachable),
+                    Box::new(AppRpcWeb3ErrorKind::ServerUnreachable),
                     &ValidationFailureClockReal::default(),
                 ),
             )))
@@ -782,7 +782,7 @@ mod tests {
             (
                 tx2.hash,
                 RecheckRequired(ValidationStatus::Reattempting(PreviousAttempts::new(
-                    Box::new(AppRpcErrorKind::ServerUnreachable),
+                    Box::new(AppRpcWeb3ErrorKind::ServerUnreachable),
                     &ValidationFailureClockReal::default(),
                 ))),
             ),
@@ -799,7 +799,7 @@ mod tests {
         assert_eq!(
             updated_txs[1].status,
             RecheckRequired(ValidationStatus::Reattempting(PreviousAttempts::new(
-                Box::new(AppRpcErrorKind::ServerUnreachable),
+                Box::new(AppRpcWeb3ErrorKind::ServerUnreachable),
                 &ValidationFailureClockReal::default()
             )))
         );
