+++ node/src/accountant/db_access_objects/payable_dao.rs
@@ -417,7 +417,7 @@ mod mark_pending_payable_associated_functions {
             },
             Err(errs) => {
                 let err_msg = format!(
-                    "Multi-row update to mark pending payable hit these errors: {:?}",
+                    "Multi-row update to mark pending payable hit these app_rpc_web3_error_kind: {:?}",
                     errs
                 );
                 Err(PayableDaoError::RusqliteError(err_msg))
@@ -874,7 +874,7 @@ mod tests {
         assert_eq!(
             result,
             Err(PayableDaoError::RusqliteError(
-                "Multi-row update to mark pending payable hit these errors: [SqliteFailure(\
+                "Multi-row update to mark pending payable hit these app_rpc_web3_error_kind: [SqliteFailure(\
                 Error { code: ReadOnly, extended_code: 8 }, Some(\"attempt to write a readonly \
                 database\"))]"
                     .to_string()
