+++ node/src/proxy_client/stream_writer.rs
@@ -104,7 +104,7 @@ impl StreamWriter {
                                 );
                                 return Err(());
                             } else {
-                                // TODO this could be exploitable and inefficient: if we keep getting non-dead-stream errors, we go into a tight loop and do not return
+                                // TODO this could be exploitable and inefficient: if we keep getting non-dead-stream app_rpc_web3_error_kind, we go into a tight loop and do not return
                                 warning!(self.logger, "Continuing after write error: {}", e);
                                 self.sequence_buffer.repush(packet);
                             }
