+++ node/src/stream_writer_sorted.rs
@@ -100,7 +100,7 @@ impl StreamWriterSorted {
                                 );
                                 return WriteBufferStatus::StreamInError;
                             } else {
-                                // TODO this could be exploitable and inefficient: if we keep getting non-dead-stream errors, we go into a tight loop and do not return
+                                // TODO this could be exploitable and inefficient: if we keep getting non-dead-stream app_rpc_web3_error_kind, we go into a tight loop and do not return
                                 warning!(self.logger, "Continuing after write error: {}", e);
                                 self.sequence_buffer.repush(packet);
                             }
